Move RenderProgram to Video
